tqec.templates.rpng -> tqec.compile.specs.library.generators._testing
